Changed scene structure
BaseHolokitScene is what should be loaded first and it should include all managers (gestures, speech, spatial mapping etc) so that there is only one manager active. Each different scene should be loaded in in addition to the base scene, adding only the specific visuals/objects of that scene.

In the base scene, the object SceneManager should have a script that handles loading/unloading of scenes. All keywords/responses needs to be defined in this scenes SpeechInputSource